+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## v0.2.0
+* Normalised interfaces and traits
+* Added Facade plugin binding mechanism
+
 ## v0.1.0
 * Created initial Facade generator structure